refactor: /test/lotto/lotto에 만들었던 test 코드 이동
